pyproject.toml
[project]
name = "xmd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "discord-py>=2.7.0",
    "fastapi>=0.134.0",
    "jinja2>=3.1.6",
    "tweety-ns>=2.4.1",
    "uvicorn>=0.41.0",
    "zipstream-ng>=1.9.0",
]

[tool.setuptools.packages.find]
where = ["."]
